tests/api_smoke_tests/requirements.txt
google-cloud-vision>=3.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
requests>=2.31.0
